// frontend-react/src/lib/contracts/ws.generated.ts
// Generated file. Do not edit.

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ClientControlMessage = {
  type: "resize";
  cols: number;
  rows: number;
  visible: boolean;
};

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ServerControlMessage =
  | {
      type: "attached";
      byte_offset: number;
      data_lost: boolean;
      cols: number;
      rows: number;
    }
  | { type: "pty_resized"; cols: number; rows: number }
  | { type: "tab_closed" };